feat: adds table component
