+++ src/app/pages/planning/planning.component.html
@@ -1,197 +1,2 @@
-<section class="min-h-[calc(100vh-80px)] bg-[#1c1c1e] text-white">
-  <div class="max-w-6xl mx-auto px-4 md:px-8 py-8">
-  <div class="flex items-center justify-between gap-4 mb-4">
-    <div class="flex items-center gap-3">
-      <button
-        type="button"
-        [matMenuTriggerFor]="categoryTypeMenu"
-        class="text-white font-semibold text-sm px-4 py-2 rounded-full flex items-center gap-2 border-none cursor-pointer transition-colors"
-        [ngClass]="{
-          'bg-red-500 hover:bg-red-400': selectedCategoryType === 'despesa',
-          'bg-[#10b780] hover:bg-[#0ea371]': selectedCategoryType === 'receita'
-        }"
-        aria-label="Selecionar tipo de categoria"
-      >
-        <span>{{ categoryTypeLabel }}</span>
-        <mat-icon class="text-white text-[20px] leading-[20px]">expand_more</mat-icon>
-      </button>
-
-      <mat-menu #categoryTypeMenu="matMenu" xPosition="after" class="!bg-[#2c2c2e]">
-        <button mat-menu-item (click)="setCategoryType('despesa')">
-          <mat-icon class="!text-red-400">trending_down</mat-icon>
-          <span class="text-[#ffffffb3]">Categoria de Despesas</span>
-        </button>
-        <button mat-menu-item (click)="setCategoryType('receita')">
-          <mat-icon class="!text-[#10b780]">trending_up</mat-icon>
-          <span class="text-[#ffffffb3]">Categoria de Receitas</span>
-        </button>
-      </mat-menu>
-    </div>
-
-    <div class="flex items-center gap-2">
-      <div *ngIf="showSearch" class="w-[240px] md:w-[320px]">
-        <input
-          matInput
-          [value]="searchValue"
-          (input)="searchValue = $any($event.target).value; applyFilter(searchValue)"
-          placeholder="Buscar..."
-          class="w-full h-10 px-3 rounded-full !text-white placeholder:!text-white/50 !bg-[#2c2c2e] border border-white/10 focus:outline-none"
-          aria-label="Buscar categorias"
-        />
-      </div>
-
-      <button
-        type="button"
-        mat-icon-button
-        (click)="onAddCategory()"
-        class="text-white hover:bg-white/10"
-        matTooltip="Adicionar"
-        aria-label="Adicionar categoria"
-      >
-        <mat-icon class="text-white">add</mat-icon>
-      </button>
-
-      <button
-        type="button"
-        mat-icon-button
-        (click)="onRefresh()"
-        class="text-white hover:bg-white/10"
-        matTooltip="Atualizar"
-        aria-label="Atualizar"
-      >
-        <mat-icon class="text-white">refresh</mat-icon>
-      </button>
-
-      <button
-        type="button"
-        mat-icon-button
-        (click)="toggleSearch()"
-        class="text-white hover:bg-white/10"
-        matTooltip="Buscar"
-        aria-label="Buscar"
-      >
-        <mat-icon class="text-white">search</mat-icon>
-      </button>
-
-      <button
-        type="button"
-        [matMenuTriggerFor]="moreMenu"
-        mat-icon-button
-        class="text-white hover:bg-white/10"
-        matTooltip="Mais"
-        aria-label="Mais ações"
-      >
-        <mat-icon class="text-white">more_vert</mat-icon>
-      </button>
-
-      <mat-menu #moreMenu="matMenu" xPosition="before" class="!bg-[#2c2c2e]">
-        <button mat-menu-item (click)="applyFilter('')">
-          <mat-icon class="!text-[#ffffffb3]">filter_alt_off</mat-icon>
-          <span class="text-[#ffffffb3]">Limpar filtro</span>
-        </button>
-      </mat-menu>
-    </div>
-  </div>
-
-  <div class="rounded-2xl overflow-hidden bg-[#2c2c2e] border border-white/10">
-    <div class="overflow-x-auto">
-      <table mat-table [dataSource]="dataSource" matSort class="w-full !bg-transparent table-fixed">
-      <ng-container matColumnDef="name">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header class="!text-white/80 !font-semibold !bg-transparent relative pr-4">
-          <span>Nome</span>
-          <span
-            appColumnResize="name"
-            class="absolute right-0 top-0 h-full w-4 cursor-col-resize"
-            aria-hidden="true"
-          >
-            <span class="absolute right-1 top-1/2 -translate-y-1/2 h-6 w-px bg-white/20 opacity-60 hover:opacity-100"></span>
-          </span>
-        </th>
-        <td mat-cell *matCellDef="let row" class="!text-white/80 !bg-transparent">{{ row.name }}</td>
-      </ng-container>
-
-      <ng-container matColumnDef="icon">
-        <th mat-header-cell *matHeaderCellDef class="!text-white/80 !font-semibold !bg-transparent relative pr-4">
-          <span>Ícone</span>
-          <span
-            appColumnResize="icon"
-            class="absolute right-0 top-0 h-full w-4 cursor-col-resize"
-            aria-hidden="true"
-          >
-            <span class="absolute right-1 top-1/2 -translate-y-1/2 h-6 w-px bg-white/20 opacity-60 hover:opacity-100"></span>
-          </span>
-        </th>
-        <td mat-cell *matCellDef="let row" class="!text-white/80 !bg-transparent">
-          <mat-icon class="!text-white/70 text-[18px] leading-[18px]">{{ row.icon }}</mat-icon>
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="color">
-        <th mat-header-cell *matHeaderCellDef class="!text-white/80 !font-semibold !bg-transparent relative pr-4">
-          <span>Cor</span>
-          <span
-            appColumnResize="color"
-            class="absolute right-0 top-0 h-full w-4 cursor-col-resize"
-            aria-hidden="true"
-          >
-            <span class="absolute right-1 top-1/2 -translate-y-1/2 h-6 w-px bg-white/20 opacity-60 hover:opacity-100"></span>
-          </span>
-        </th>
-        <td mat-cell *matCellDef="let row" class="!bg-transparent">
-          <span class="inline-block w-3.5 h-3.5 rounded-full" [style.background]="row.color"></span>
-        </td>
-      </ng-container>
-
-      <ng-container matColumnDef="actions">
-        <th mat-header-cell *matHeaderCellDef class="!text-white/80 !font-semibold text-right !bg-transparent">Ações</th>
-        <td mat-cell *matCellDef="let row" class="text-right !bg-transparent">
-          <button
-            type="button"
-            mat-icon-button
-            class="text-white/80 hover:bg-white/10 !w-8 !h-8 !p-0"
-            matTooltip="Detalhes"
-            (click)="onRowAction('details', row)"
-            aria-label="Detalhes"
-          >
-            <mat-icon class="text-white/70 text-[18px] leading-[18px]">description</mat-icon>
-          </button>
-          <button
-            type="button"
-            mat-icon-button
-            class="text-white/80 hover:bg-white/10 !w-8 !h-8 !p-0"
-            matTooltip="Editar"
-            (click)="onRowAction('edit', row)"
-            aria-label="Editar"
-          >
-            <mat-icon class="text-white/70 text-[18px] leading-[18px]">edit</mat-icon>
-          </button>
-          <button
-            type="button"
-            mat-icon-button
-            class="text-white/80 hover:bg-white/10 !w-8 !h-8 !p-0"
-            matTooltip="Arquivar"
-            (click)="onRowAction('archive', row)"
-            aria-label="Arquivar"
-          >
-            <mat-icon class="text-white/70 text-[18px] leading-[18px]">inventory_2</mat-icon>
-          </button>
-        </td>
-      </ng-container>
-
-      <tr mat-header-row *matHeaderRowDef="displayedColumns" class="!bg-white/5"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns" class="!border-b !border-white/10 hover:!bg-white/5"></tr>
-      </table>
-    </div>
-
-    <div class="border-t border-white/10">
-      <mat-paginator
-        class="!bg-[#2c2c2e]"
-        [pageSize]="50"
-        [pageSizeOptions]="[10, 25, 50, 100]"
-        showFirstLastButtons
-      ></mat-paginator>
-    </div>
-  </div>
-  </div>
-</section>
+<app-table></app-table>
